fix COOTensor.add format

+++ mindspore/python/mindspore/common/sparse_tensor.py
@@ -535,9 +535,11 @@ class COOTensor(COOTensor_):
             >>> thres = Tensor(0, dtype=mstype.int32)
             >>> out = input0.add(input1, thres)
             >>> print(out)
-            COOTensor(shape = [3, 4], dtype = Int32, indices=Tensor(shape=[4, 2],
-            dtype = Int64, value=[[0 0], [0 1], [1 1], [1 2]]),  values=Tensor(shape[4],
-            dtype=Int32, value=[3 1 4 2]))
+            COOTensor(shape=[3, 4], dtype=Int32, indices=Tensor(shape=[4, 2], dtype=Int64, value=
+            [[0 0]
+            [0 1]
+            [1 1]
+            [1 2]]), values=Tensor(shape[4], dtype=Int32, value=[3 1 4 2]))
         """
         return tensor_operator_registry.get('coo_add')(self, other, thresh)
